+++ configs/pie-code-connect/pie-tag.figma.batch.js
@@ -38,7 +38,7 @@ const label = getInstanceProp('getString', '[𝐓] Label');
 // Get icon instance and add the slot prop to the snippet
 const hasIcon = isIconOnly || hasLeadingIcon;
 const iconInstance = hasIcon && isIconOnly ? getInstanceProp('getInstanceSwap', 'Icon') : getInstanceProp('getInstanceSwap', 'Replace leading icon');
-const iconSnippet = hasIcon && getIconSnippet(iconInstance, (code) => code.replace('></', ' slot="icon"></'));
+const iconSnippet = hasIcon && getIconSnippet(iconInstance, 'icon');
 
 const props = [
     renderProp('variant', variant, 'neutral'),
@@ -58,5 +58,5 @@ const template = figma.code`<${selectedComponentName}
 export default {
     example: template,
     imports: [getImportStatement(componentName, componentNameReact)],
-    id: 'pie-tag',
+    id: componentName,
 };